Update header file

+++ Inkpad-Core/Model/WDBezierSegment.h
@@ -39,7 +39,7 @@ BOOL WDBezierSegmentIntersectsRect(WDBezierSegment seg, CGRect rect);
 BOOL WDLineInRect(CGPoint a, CGPoint b, CGRect test);
 
 BOOL WDBezierSegmentIsStraight(WDBezierSegment segment);
-BOOL WDBezierSegmentIsFlat(WDBezierSegment seg, float tolerance);
+BOOL WDBezierSegmentIsFlat(WDBezierSegment seg, CGFloat tolerance);
 void WDBezierSegmentFlatten(WDBezierSegment seg, CGPoint **vertices, NSUInteger *size, NSUInteger *index);
 CGPoint WDBezierSegmentSplit(WDBezierSegment seg, WDBezierSegment *L, WDBezierSegment *R);
 CGPoint WDBezierSegmentSplitAtT(WDBezierSegment seg, WDBezierSegment *L, WDBezierSegment *R, float t);
